Move public APIs that do unsafe operations into `unsafe/` subdirs
We should make them to use less-convenient
