+++ docs/public_api_list.md
@@ -108,8 +108,6 @@ Basic APIs are exported here:
 - [`option-t/plain_option/and`](../packages/option-t/src/plain_option/and.ts)
 - [`option-t/plain_option/and_then`](../packages/option-t/src/plain_option/and_then.ts)
 - [`option-t/plain_option/and_then_async`](../packages/option-t/src/plain_option/and_then_async.ts)
-- [`option-t/plain_option/as_mut`](../packages/option-t/src/plain_option/as_mut.ts)
-- [`option-t/plain_option/drop`](../packages/option-t/src/plain_option/drop.ts)
 - [`option-t/plain_option/equal`](../packages/option-t/src/plain_option/equal.ts)
 - [`option-t/plain_option/filter`](../packages/option-t/src/plain_option/filter.ts)
 - [`option-t/plain_option/flatten`](../packages/option-t/src/plain_option/flatten.ts)
@@ -129,6 +127,8 @@ Basic APIs are exported here:
 - [`option-t/plain_option/to_nullable`](../packages/option-t/src/plain_option/to_nullable.ts)
 - [`option-t/plain_option/to_undefinable`](../packages/option-t/src/plain_option/to_undefinable.ts)
 - [`option-t/plain_option/transpose`](../packages/option-t/src/plain_option/transpose.ts)
+- [`option-t/plain_option/unsafe/as_mut`](../packages/option-t/src/plain_option/unsafe/as_mut.ts)
+- [`option-t/plain_option/unsafe/drop`](../packages/option-t/src/plain_option/unsafe/drop.ts)
 - [`option-t/plain_option/unwrap_or`](../packages/option-t/src/plain_option/unwrap_or.ts)
 - [`option-t/plain_option/unwrap_or_else`](../packages/option-t/src/plain_option/unwrap_or_else.ts)
 - [`option-t/plain_option/unwrap_or_else_async`](../packages/option-t/src/plain_option/unwrap_or_else_async.ts)
@@ -153,12 +153,10 @@ Basic APIs are exported here:
 - [`option-t/plain_result/and`](../packages/option-t/src/plain_result/and.ts)
 - [`option-t/plain_result/and_then`](../packages/option-t/src/plain_result/and_then.ts)
 - [`option-t/plain_result/and_then_async`](../packages/option-t/src/plain_result/and_then_async.ts)
-- [`option-t/plain_result/as_mut`](../packages/option-t/src/plain_result/as_mut.ts)
 - [`option-t/plain_result/deprecated/try_catch_with_assert_error`](../packages/option-t/src/plain_result/deprecated/try_catch_with_assert_error.ts) (__deprecated__)
 - [`option-t/plain_result/deprecated/try_catch_with_assert_error_async`](../packages/option-t/src/plain_result/deprecated/try_catch_with_assert_error_async.ts) (__deprecated__)
 - [`option-t/plain_result/deprecated/unwrap_or_throw_error`](../packages/option-t/src/plain_result/deprecated/unwrap_or_throw_error.ts) (__deprecated__. Use `option-t/plain_result/unwrap_or_throw` instead.)
 - [`option-t/plain_result/deprecated/unwrap_or_throw_unknown`](../packages/option-t/src/plain_result/deprecated/unwrap_or_throw_unknown.ts) (__deprecated__. Use `option-t/plain_result/unwrap_or_throw` instead.)
-- [`option-t/plain_result/drop`](../packages/option-t/src/plain_result/drop.ts)
 - [`option-t/plain_result/equal`](../packages/option-t/src/plain_result/equal.ts)
 - [`option-t/plain_result/flatten`](../packages/option-t/src/plain_result/flatten.ts)
 - [`option-t/plain_result/from_promise_settled_result`](../packages/option-t/src/plain_result/from_promise_settled_result.ts)
@@ -181,6 +179,8 @@ Basic APIs are exported here:
 - [`option-t/plain_result/transpose`](../packages/option-t/src/plain_result/transpose.ts)
 - [`option-t/plain_result/try_catch`](../packages/option-t/src/plain_result/try_catch.ts)
 - [`option-t/plain_result/try_catch_async`](../packages/option-t/src/plain_result/try_catch_async.ts)
+- [`option-t/plain_result/unsafe/as_mut`](../packages/option-t/src/plain_result/unsafe/as_mut.ts)
+- [`option-t/plain_result/unsafe/drop`](../packages/option-t/src/plain_result/unsafe/drop.ts)
 - [`option-t/plain_result/unwrap_or`](../packages/option-t/src/plain_result/unwrap_or.ts)
 - [`option-t/plain_result/unwrap_or_else`](../packages/option-t/src/plain_result/unwrap_or_else.ts)
 - [`option-t/plain_result/unwrap_or_else_async`](../packages/option-t/src/plain_result/unwrap_or_else_async.ts)
