+++ packages/api_tests/__tests__/exposed_path/plain_option.test.mjs
@@ -10,8 +10,6 @@ testImportExposedPath(
         './plain_option/and': null,
         './plain_option/and_then': null,
         './plain_option/and_then_async': null,
-        './plain_option/as_mut': null,
-        './plain_option/drop': null,
         './plain_option/equal': null,
         './plain_option/filter': null,
         './plain_option/flatten': null,
@@ -33,6 +31,8 @@ testImportExposedPath(
         './plain_option/to_nullable': null,
         './plain_option/to_undefinable': null,
         './plain_option/transpose': null,
+        './plain_option/unsafe/as_mut': null,
+        './plain_option/unsafe/drop': null,
         './plain_option/unwrap_or': null,
         './plain_option/unwrap_or_else': null,
         './plain_option/unwrap_or_else_async': null,
