+++ packages/api_tests/__tests__/exposed_path/plain_result.test.mjs
@@ -10,12 +10,10 @@ testImportExposedPath(
         './plain_result/and': null,
         './plain_result/and_then': null,
         './plain_result/and_then_async': null,
-        './plain_result/as_mut': null,
         './plain_result/deprecated/try_catch_with_assert_error': null,
         './plain_result/deprecated/try_catch_with_assert_error_async': null,
         './plain_result/deprecated/unwrap_or_throw_error': null,
         './plain_result/deprecated/unwrap_or_throw_unknown': null,
-        './plain_result/drop': null,
         './plain_result/equal': null,
         './plain_result/flatten': null,
         './plain_result/from_promise_settled_result': null,
@@ -40,6 +38,8 @@ testImportExposedPath(
         './plain_result/transpose': null,
         './plain_result/try_catch': null,
         './plain_result/try_catch_async': null,
+        './plain_result/unsafe/as_mut': null,
+        './plain_result/unsafe/drop': null,
         './plain_result/unwrap_or': null,
         './plain_result/unwrap_or_else': null,
         './plain_result/unwrap_or_else_async': null,
