+++ packages/api_tests/__tests__/plain_result/as_mut.test.mjs
@@ -1,7 +1,7 @@
 import test from 'ava';
 
-import { asMutResult } from 'option-t/plain_result/as_mut';
 import { createOk, createErr } from 'option-t/plain_result/result';
+import { asMutResult } from 'option-t/plain_result/unsafe/as_mut';
 
 const FUNC_LIST = [createOk, createErr];
