+++ packages/api_tests/__tests__/plain_result/drop/test_drop_ok.test.mjs
@@ -1,8 +1,8 @@
 /* eslint-disable no-param-reassign */
 import test from 'ava';
 
-import { unsafeDropOkForResult } from 'option-t/plain_result/drop';
 import { createOk, createErr, isOk } from 'option-t/plain_result/result';
+import { unsafeDropOkForResult } from 'option-t/plain_result/unsafe/drop';
 
 test('with Ok', (t) => {
     const expected = Symbol('');
