+++ packages/option-t/package.json
@@ -1271,42 +1271,6 @@
                 "default": "./esm/plain_option/and_then_async.js"
             }
         },
-        "./plain_option/as_mut": {
-            "import": {
-                "types": "./esm/plain_option/as_mut.d.ts",
-                "default": "./esm/plain_option/as_mut.js"
-            },
-            "module-sync": {
-                "types": "./esm/plain_option/as_mut.d.ts",
-                "default": "./esm/plain_option/as_mut.js"
-            },
-            "require": {
-                "types": "./cjs/plain_option/as_mut.d.cts",
-                "default": "./cjs/plain_option/as_mut.cjs"
-            },
-            "default": {
-                "types": "./esm/plain_option/as_mut.d.ts",
-                "default": "./esm/plain_option/as_mut.js"
-            }
-        },
-        "./plain_option/drop": {
-            "import": {
-                "types": "./esm/plain_option/drop.d.ts",
-                "default": "./esm/plain_option/drop.js"
-            },
-            "module-sync": {
-                "types": "./esm/plain_option/drop.d.ts",
-                "default": "./esm/plain_option/drop.js"
-            },
-            "require": {
-                "types": "./cjs/plain_option/drop.d.cts",
-                "default": "./cjs/plain_option/drop.cjs"
-            },
-            "default": {
-                "types": "./esm/plain_option/drop.d.ts",
-                "default": "./esm/plain_option/drop.js"
-            }
-        },
         "./plain_option/equal": {
             "import": {
                 "types": "./esm/plain_option/equal.d.ts",
@@ -1685,6 +1649,42 @@
                 "default": "./esm/plain_option/transpose.js"
             }
         },
+        "./plain_option/unsafe/as_mut": {
+            "import": {
+                "types": "./esm/plain_option/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_option/unsafe/as_mut.js"
+            },
+            "module-sync": {
+                "types": "./esm/plain_option/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_option/unsafe/as_mut.js"
+            },
+            "require": {
+                "types": "./cjs/plain_option/unsafe/as_mut.d.cts",
+                "default": "./cjs/plain_option/unsafe/as_mut.cjs"
+            },
+            "default": {
+                "types": "./esm/plain_option/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_option/unsafe/as_mut.js"
+            }
+        },
+        "./plain_option/unsafe/drop": {
+            "import": {
+                "types": "./esm/plain_option/unsafe/drop.d.ts",
+                "default": "./esm/plain_option/unsafe/drop.js"
+            },
+            "module-sync": {
+                "types": "./esm/plain_option/unsafe/drop.d.ts",
+                "default": "./esm/plain_option/unsafe/drop.js"
+            },
+            "require": {
+                "types": "./cjs/plain_option/unsafe/drop.d.cts",
+                "default": "./cjs/plain_option/unsafe/drop.cjs"
+            },
+            "default": {
+                "types": "./esm/plain_option/unsafe/drop.d.ts",
+                "default": "./esm/plain_option/unsafe/drop.js"
+            }
+        },
         "./plain_option/unwrap_or": {
             "import": {
                 "types": "./esm/plain_option/unwrap_or.d.ts",
@@ -1829,24 +1829,6 @@
                 "default": "./esm/plain_result/and_then_async.js"
             }
         },
-        "./plain_result/as_mut": {
-            "import": {
-                "types": "./esm/plain_result/as_mut.d.ts",
-                "default": "./esm/plain_result/as_mut.js"
-            },
-            "module-sync": {
-                "types": "./esm/plain_result/as_mut.d.ts",
-                "default": "./esm/plain_result/as_mut.js"
-            },
-            "require": {
-                "types": "./cjs/plain_result/as_mut.d.cts",
-                "default": "./cjs/plain_result/as_mut.cjs"
-            },
-            "default": {
-                "types": "./esm/plain_result/as_mut.d.ts",
-                "default": "./esm/plain_result/as_mut.js"
-            }
-        },
         "./plain_result/deprecated/try_catch_with_assert_error": {
             "import": {
                 "types": "./esm/plain_result/deprecated/try_catch_with_assert_error.d.ts",
@@ -1917,24 +1899,6 @@
             "default": {
                 "types": "./esm/plain_result/deprecated/unwrap_or_throw_unknown.d.ts",
                 "default": "./esm/plain_result/deprecated/unwrap_or_throw_unknown.js"
-            }
-        },
-        "./plain_result/drop": {
-            "import": {
-                "types": "./esm/plain_result/drop.d.ts",
-                "default": "./esm/plain_result/drop.js"
-            },
-            "module-sync": {
-                "types": "./esm/plain_result/drop.d.ts",
-                "default": "./esm/plain_result/drop.js"
-            },
-            "require": {
-                "types": "./cjs/plain_result/drop.d.cts",
-                "default": "./cjs/plain_result/drop.cjs"
-            },
-            "default": {
-                "types": "./esm/plain_result/drop.d.ts",
-                "default": "./esm/plain_result/drop.js"
             }
         },
         "./plain_result/equal": {
@@ -2367,6 +2331,42 @@
             "default": {
                 "types": "./esm/plain_result/try_catch_async.d.ts",
                 "default": "./esm/plain_result/try_catch_async.js"
+            }
+        },
+        "./plain_result/unsafe/as_mut": {
+            "import": {
+                "types": "./esm/plain_result/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_result/unsafe/as_mut.js"
+            },
+            "module-sync": {
+                "types": "./esm/plain_result/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_result/unsafe/as_mut.js"
+            },
+            "require": {
+                "types": "./cjs/plain_result/unsafe/as_mut.d.cts",
+                "default": "./cjs/plain_result/unsafe/as_mut.cjs"
+            },
+            "default": {
+                "types": "./esm/plain_result/unsafe/as_mut.d.ts",
+                "default": "./esm/plain_result/unsafe/as_mut.js"
+            }
+        },
+        "./plain_result/unsafe/drop": {
+            "import": {
+                "types": "./esm/plain_result/unsafe/drop.d.ts",
+                "default": "./esm/plain_result/unsafe/drop.js"
+            },
+            "module-sync": {
+                "types": "./esm/plain_result/unsafe/drop.d.ts",
+                "default": "./esm/plain_result/unsafe/drop.js"
+            },
+            "require": {
+                "types": "./cjs/plain_result/unsafe/drop.d.cts",
+                "default": "./cjs/plain_result/unsafe/drop.cjs"
+            },
+            "default": {
+                "types": "./esm/plain_result/unsafe/drop.d.ts",
+                "default": "./esm/plain_result/unsafe/drop.js"
             }
         },
         "./plain_result/unwrap_or": {
