+++ packages/option-t/src/plain_option/index.ts
@@ -68,8 +68,8 @@ export { unwrapOrElseAsyncForOption } from './unwrap_or_else_async.js';
 // XXX:
 //  We don't expose these itens that is unsafe operation.
 //
-//  - as_mut
-//  - drop
+//  - unsafe/as_mut
+//  - unsafe/drop
 //
 // XXX:
 //  _equals, we don't expose it by this due to that is provided for exception case
