+++ packages/option-t/src/plain_result/unsafe/as_mut.ts
@@ -1,6 +1,6 @@
-import { assertIsFrozen } from '../internal/assert.js';
-import type { MutResult as InternalMutResult } from './internal/mutable.js';
-import type { Result } from './result.js';
+import { assertIsFrozen } from '../../internal/assert.js';
+import type { MutResult as InternalMutResult } from '../internal/mutable.js';
+import type { Result } from '../result.js';
 
 /**
  *  This allows to mutate the value to save needless allocation.
