+++ packages/test_module_resolution_bundler/src/plain_result.cts
@@ -2,8 +2,6 @@ import {} from 'option-t/plain_result';
 import {} from 'option-t/plain_result/and';
 import {} from 'option-t/plain_result/and_then';
 import {} from 'option-t/plain_result/and_then_async';
-import {} from 'option-t/plain_result/as_mut';
-import {} from 'option-t/plain_result/drop';
 import {} from 'option-t/plain_result/equal';
 import {} from 'option-t/plain_result/flatten';
 import {} from 'option-t/plain_result/from_promise_settled_result';
@@ -28,6 +26,8 @@ import {} from 'option-t/plain_result/to_undefinable';
 import {} from 'option-t/plain_result/try_catch';
 import {} from 'option-t/plain_result/try_catch_async';
 import {} from 'option-t/plain_result/transpose';
+import {} from 'option-t/plain_result/unsafe/as_mut';
+import {} from 'option-t/plain_result/unsafe/drop';
 import {} from 'option-t/plain_result/unwrap_or';
 import {} from 'option-t/plain_result/unwrap_or_else';
 import {} from 'option-t/plain_result/unwrap_or_else_async';
