+++ packages/test_module_resolution_node16/src/plain_option.mts
@@ -2,8 +2,6 @@ import {} from 'option-t/plain_option';
 import {} from 'option-t/plain_option/and';
 import {} from 'option-t/plain_option/and_then';
 import {} from 'option-t/plain_option/and_then_async';
-import {} from 'option-t/plain_option/as_mut';
-import {} from 'option-t/plain_option/drop';
 import {} from 'option-t/plain_option/equal';
 import {} from 'option-t/plain_option/filter';
 import {} from 'option-t/plain_option/flatten';
@@ -25,6 +23,8 @@ import {} from 'option-t/plain_option/or_else_async';
 import {} from 'option-t/plain_option/transpose';
 import {} from 'option-t/plain_option/to_nullable';
 import {} from 'option-t/plain_option/to_undefinable';
+import {} from 'option-t/plain_option/unsafe/as_mut';
+import {} from 'option-t/plain_option/unsafe/drop';
 import {} from 'option-t/plain_option/unwrap_or';
 import {} from 'option-t/plain_option/unwrap_or_else';
 import {} from 'option-t/plain_option/unwrap_or_else_async';
